moved svg icons to assets folder
